backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==1.10.12
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pymongo==4.4.0
dnspython==2.4.2
python-multipart==0.0.6
sentence-transformers==2.2.2
spacy==3.5.3
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl

scikit-learn==1.3.0
pyyaml==6.0
#rasa==3.7.3   #Only if you intend to train with Rasa locally
